isTakeEssentialNutrients 커스텀 훅 추가 및 컨테이너 padding bottom 여유공간 추가
