+++ components/layout/ContainerWithBottomNav.tsx
@@ -13,7 +13,8 @@ function ContainerWithBottomNav({ children, headerHeight = '' }: Props) {
     <>
       {/* margin이 더 적합한 값일 수 있으나 스크롤 바닥에서는 하단 margin이 안들어감 이러한 이유로 padding 사용. (이런 이유로 다른 사이트들도 padding 쓰는듯) */}
       {/* 때문에 여기서 background color등 사용 X */}
-      <div className={'w-full pb-14' + ` ${headerHeight}`}>
+      {/* pb-14가 적당하나 pb-20으로 여유를 둠 */}
+      <div className={'w-full pb-20' + ` ${headerHeight}`}>
         {children}
       </div>
       <BottomNavBar />
